requirements.txt
streamlit==1.33.0
seaborn
pandas
matplotlib.pyplot
